+++ alibi_detect/od/ae.py
@@ -52,8 +52,9 @@ class OutlierAE(BaseDetector, FitMixin, ThresholdMixin):
                             'or `encoder_net`, `decoder_net` (tf.keras.Sequential).')
 
         # set metadata
-        self.meta['detector_type'] = 'offline'
+        self.meta['detector_type'] = 'outlier'
         self.meta['data_type'] = data_type
+        self.meta['online'] = False
 
     def fit(self,
             X: np.ndarray,