+++ alibi_detect/od/seq2seq.py
@@ -83,8 +83,9 @@ class OutlierSeq2Seq(BaseDetector, FitMixin, ThresholdMixin):
                             '`latent_dim` (int) or `threshold_net` (tf.keras.Sequential)')
 
         # set metadata
-        self.meta['detector_type'] = 'offline'
+        self.meta['detector_type'] = 'outlier'
         self.meta['data_type'] = 'time-series'
+        self.meta['online'] = False
 
     def fit(self,
             X: np.ndarray,